requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
exceptiongroup==1.3.0
idna==3.11
iniconfig==2.1.0
jsonpath==0.82
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
py==1.11.0
Pygments==2.19.2
pytest==8.4.2
pytest-dependency==0.5.1
pytest-html==3.2.0
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-stats==1.0.1
PyYAML==6.0.1
requests==2.31.0
tomli==2.3.0
typing_extensions==4.15.0
urllib3==1.26.6
